+++ components/Collections.vue
@@ -6,8 +6,8 @@
 			<template #details="slot">
 				<Collection :collection="slot.item" :mapOptions="mapOptions">
 					<template #title><span class="hidden" /></template>
-					<template #before-description><slot name="collection-before-description" /></template>
-					<template #end><slot name="collection-end" /></template>
+					<template #before-description="scope"><slot name="collection-before-description" v-bind="scope" /></template>
+					<template #end="scope"><slot name="collection-end" v-bind="scope" /></template>
 					<template #spatial-extents="scope"><slot name="collection-spatial-extents" v-bind="scope" /></template>
 					<template #temporal-extents="scope"><slot name="collection-temporal-extents" v-bind="scope" /></template>
 				</Collection>