+++ components/FileFormats.vue
@@ -19,8 +19,8 @@
 				<FileFormat :id="slot.summary.identifier" :format="slot.item" :type="slot.item.type">
 					<template #title><span class="hidden" /></template>
 					<template #badges><span class="hidden" /></template>
-					<template #before-description><slot name="file-format-before-description" /></template>
-					<template #end><slot name="file-format-end" /></template>
+					<template #before-description="scope"><slot name="file-format-before-description" v-bind="scope" /></template>
+					<template #end="scope"><slot name="file-format-end" v-bind="scope" /></template>
 				</FileFormat>
 			</template>
 		</SearchableList>