+++ components/UdfRuntimes.vue
@@ -20,8 +20,8 @@
 			<template #details="slot">
 				<UdfRuntime :id="slot.summary.identifier" :runtime="slot.item">
 					<template #title><span class="hidden" /></template>
-					<template #badges><slot name="udf-runtime-badges" /></template>
-					<template #before-description><slot name="udf-runtime-before-description" /></template>
+					<template #badges="scope"><slot name="udf-runtime-badges" v-bind="scope" /></template>
+					<template #before-description="scope"><slot name="udf-runtime-before-description" v-bind="scope" /></template>
 				</UdfRuntime>
 			</template>
 		</SearchableList>